backend/requirements.txt
google-genai
requests
httpx
python-dotenv
google-auth-oauthlib
google-api-python-client
Pillow
huggingface-hub
elevenlabs
mutagen
